src/timeout: only update nodes owned by the current user
In order to avoid hitting issues with API permissions when updating
nodes that aren't owned by the current user, even with an admin API
token, ignore nodes that have a different owner to the current user
running the timeout service.

This is most likely a temporary measure until API permissions have
been fixed as the timeout service should in principle be able to
transition all nodes from all users.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>